Add AI option to switch to mon immune to last move
Also improved defensive mon selection, which now takes into account
the known move types of the opponent too, so no switching to a
grass mon if the electric type opponent knows a fire move.
Replaced gLastLandedMoves[user] with gLastMoves[attacker] since
gLastLandedMoves is cleared if the user faints. There's still an issue
with gLastMoves not updating if it was a one hit ko.